move uuid tests from bson tests
