+++ src/tests/modules/bson.rs
@@ -18,11 +18,6 @@ use crate::{
 };
 use serde_json::{json, Value};
 
-#[cfg(feature = "uuid-0_8")]
-use crate::uuid::UuidRepresentation;
-#[cfg(feature = "uuid-0_8")]
-use uuid::Uuid;
-
 #[test]
 fn to_json() {
     let _guard = LOCK.run_concurrently();
@@ -395,147 +390,4 @@ fn debug_print() {
 
     assert_eq!(format!("{:?}", doc), normal_print);
     assert_eq!(format!("{:#?}", doc), pretty_print);
-}
-
-#[cfg(feature = "uuid-0_8")]
-#[test]
-fn test_binary_constructors() {
-    let uuid = crate::Uuid(Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap());
-    let bin = Binary::from_uuid(uuid);
-    assert_eq!(bin.bytes, uuid.0.as_bytes());
-    assert_eq!(bin.subtype, BinarySubtype::Uuid);
-
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::Standard);
-    assert_eq!(bin.bytes, uuid.0.as_bytes());
-    assert_eq!(bin.subtype, BinarySubtype::Uuid);
-
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::JavaLegacy);
-    assert_eq!(
-        bin.bytes,
-        Uuid::parse_str("7766554433221100FFEEDDCCBBAA9988")
-            .unwrap()
-            .as_bytes()
-    );
-    assert_eq!(bin.subtype, BinarySubtype::UuidOld);
-
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::CSharpLegacy);
-    assert_eq!(
-        bin.bytes,
-        Uuid::parse_str("33221100554477668899AABBCCDDEEFF")
-            .unwrap()
-            .as_bytes()
-    );
-    assert_eq!(bin.subtype, BinarySubtype::UuidOld);
-
-    // Same byte ordering as standard representation
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::PythonLegacy);
-    assert_eq!(
-        bin.bytes,
-        Uuid::parse_str("00112233445566778899AABBCCDDEEFF")
-            .unwrap()
-            .as_bytes()
-    );
-    assert_eq!(bin.subtype, BinarySubtype::UuidOld);
-}
-
-#[cfg(feature = "uuid-0_8")]
-#[test]
-fn test_binary_to_uuid_standard_rep() {
-    let uuid = crate::Uuid(Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap());
-    let bin = Binary::from_uuid(uuid);
-
-    assert_eq!(bin.to_uuid().unwrap(), uuid);
-    assert_eq!(
-        bin.to_uuid_with_representation(UuidRepresentation::Standard)
-            .unwrap(),
-        uuid
-    );
-
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::CSharpLegacy)
-        .is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
-        .is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
-        .is_err());
-}
-
-#[cfg(feature = "uuid-0_8")]
-#[test]
-fn test_binary_to_uuid_explicitly_standard_rep() {
-    let uuid = crate::Uuid(Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap());
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::Standard);
-
-    assert_eq!(bin.to_uuid().unwrap(), uuid);
-    assert_eq!(
-        bin.to_uuid_with_representation(UuidRepresentation::Standard)
-            .unwrap(),
-        uuid
-    );
-
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::CSharpLegacy)
-        .is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
-        .is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
-        .is_err());
-}
-
-#[cfg(feature = "uuid-0_8")]
-#[test]
-fn test_binary_to_uuid_java_rep() {
-    let uuid = crate::Uuid(Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap());
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::JavaLegacy);
-
-    assert!(bin.to_uuid().is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::Standard)
-        .is_err());
-
-    assert_eq!(
-        bin.to_uuid_with_representation(UuidRepresentation::JavaLegacy)
-            .unwrap(),
-        uuid
-    );
-}
-
-#[cfg(feature = "uuid-0_8")]
-#[test]
-fn test_binary_to_uuid_csharp_legacy_rep() {
-    let uuid = crate::Uuid(Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap());
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::CSharpLegacy);
-
-    assert!(bin.to_uuid().is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::Standard)
-        .is_err());
-
-    assert_eq!(
-        bin.to_uuid_with_representation(UuidRepresentation::CSharpLegacy)
-            .unwrap(),
-        uuid
-    );
-}
-
-#[cfg(feature = "uuid-0_8")]
-#[test]
-fn test_binary_to_uuid_python_legacy_rep() {
-    let uuid = crate::Uuid(Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap());
-    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::PythonLegacy);
-
-    assert!(bin.to_uuid().is_err());
-    assert!(bin
-        .to_uuid_with_representation(UuidRepresentation::Standard)
-        .is_err());
-
-    assert_eq!(
-        bin.to_uuid_with_representation(UuidRepresentation::PythonLegacy)
-            .unwrap(),
-        uuid
-    );
 }
