+++ src/uuid/test.rs
@@ -1,4 +1,10 @@
-use crate::{spec::BinarySubtype, uuid::Uuid, Binary, Bson, Document};
+use crate::{
+    spec::BinarySubtype,
+    uuid::{Uuid, UuidRepresentation},
+    Binary,
+    Bson,
+    Document,
+};
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 
@@ -13,7 +19,7 @@ fn into_bson() {
 
     let bson: Bson = uuid.into();
     let binary = Binary {
-        bytes: uuid.as_bytes().to_vec(),
+        bytes: uuid.bytes().to_vec(),
         subtype: BinarySubtype::Uuid,
     };
 
@@ -52,4 +58,157 @@ fn json() {
 
     let u_roundtrip_json: U = serde_json::from_value(json).unwrap();
     assert_eq!(u_roundtrip_json, u);
+}
+
+#[test]
+fn test_binary_constructors() {
+    let uuid = crate::Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap();
+    let bin = Binary::from_uuid(uuid);
+    assert_eq!(bin.bytes, uuid.bytes());
+    assert_eq!(bin.subtype, BinarySubtype::Uuid);
+
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::Standard);
+    assert_eq!(bin.bytes, uuid.bytes());
+    assert_eq!(bin.subtype, BinarySubtype::Uuid);
+
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::JavaLegacy);
+    assert_eq!(
+        bin.bytes,
+        Uuid::parse_str("7766554433221100FFEEDDCCBBAA9988")
+            .unwrap()
+            .bytes()
+    );
+    assert_eq!(bin.subtype, BinarySubtype::UuidOld);
+
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::CSharpLegacy);
+    assert_eq!(
+        bin.bytes,
+        Uuid::parse_str("33221100554477668899AABBCCDDEEFF")
+            .unwrap()
+            .bytes()
+    );
+    assert_eq!(bin.subtype, BinarySubtype::UuidOld);
+
+    // Same byte ordering as standard representation
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::PythonLegacy);
+    assert_eq!(
+        bin.bytes,
+        Uuid::parse_str("00112233445566778899AABBCCDDEEFF")
+            .unwrap()
+            .bytes()
+    );
+    assert_eq!(bin.subtype, BinarySubtype::UuidOld);
+}
+
+#[test]
+fn test_binary_to_uuid_standard_rep() {
+    let uuid = crate::Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap();
+    let bin = Binary::from_uuid(uuid);
+
+    assert_eq!(bin.to_uuid().unwrap(), uuid);
+    assert_eq!(
+        bin.to_uuid_with_representation(UuidRepresentation::Standard)
+            .unwrap(),
+        uuid
+    );
+
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::CSharpLegacy)
+        .is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
+        .is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
+        .is_err());
+}
+
+#[test]
+fn test_binary_to_uuid_explicitly_standard_rep() {
+    let uuid = crate::Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap();
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::Standard);
+
+    assert_eq!(bin.to_uuid().unwrap(), uuid);
+    assert_eq!(
+        bin.to_uuid_with_representation(UuidRepresentation::Standard)
+            .unwrap(),
+        uuid
+    );
+
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::CSharpLegacy)
+        .is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
+        .is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::PythonLegacy)
+        .is_err());
+}
+
+#[test]
+fn test_binary_to_uuid_java_rep() {
+    let uuid = crate::Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap();
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::JavaLegacy);
+
+    assert!(bin.to_uuid().is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::Standard)
+        .is_err());
+
+    assert_eq!(
+        bin.to_uuid_with_representation(UuidRepresentation::JavaLegacy)
+            .unwrap(),
+        uuid
+    );
+}
+
+#[test]
+fn test_binary_to_uuid_csharp_legacy_rep() {
+    let uuid = crate::Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap();
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::CSharpLegacy);
+
+    assert!(bin.to_uuid().is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::Standard)
+        .is_err());
+
+    assert_eq!(
+        bin.to_uuid_with_representation(UuidRepresentation::CSharpLegacy)
+            .unwrap(),
+        uuid
+    );
+}
+
+#[test]
+fn test_binary_to_uuid_python_legacy_rep() {
+    let uuid = crate::Uuid::parse_str("00112233445566778899AABBCCDDEEFF").unwrap();
+    let bin = Binary::from_uuid_with_representation(uuid, UuidRepresentation::PythonLegacy);
+
+    assert!(bin.to_uuid().is_err());
+    assert!(bin
+        .to_uuid_with_representation(UuidRepresentation::Standard)
+        .is_err());
+
+    assert_eq!(
+        bin.to_uuid_with_representation(UuidRepresentation::PythonLegacy)
+            .unwrap(),
+        uuid
+    );
+}
+
+#[cfg(feature = "uuid-0_8")]
+#[test]
+fn interop() {
+    let uuid = crate::Uuid::new();
+    let uuid_uuid = uuid.to_uuid_0_8();
+    assert_eq!(uuid.to_string(), uuid_uuid.to_string());
+    assert_eq!(&uuid.bytes(), uuid_uuid.as_bytes());
+
+    let back: crate::Uuid = uuid_uuid.into();
+    assert_eq!(back, uuid);
+
+    let d_bson = doc! { "uuid": uuid };
+    let d_uuid = doc! { "uuid": uuid_uuid };
+    assert_eq!(d_bson, d_uuid);
 }
